Added basic security

+++ src/main/java/dennis/novi/livelyEvents/controller/UserController.java
@@ -1,5 +1,6 @@
 package dennis.novi.livelyEvents.controller;
 
+import dennis.novi.livelyEvents.exception.NotAuthorizedException;
 import dennis.novi.livelyEvents.exception.RecordNotFoundException;
 import dennis.novi.livelyEvents.model.User;
 import dennis.novi.livelyEvents.model.Venue;
@@ -14,11 +15,13 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
 @RestController
+@CrossOrigin(origins = "http://localhost:3000", maxAge=3600)
 
 public class UserController {
     @Autowired
@@ -29,25 +32,17 @@ public class UserController {
         List<User> users = userService.getAllUsers();
         return new ResponseEntity<>(users, HttpStatus.OK);
     }
-    @GetMapping(value = "/users/username/{username}")
-    public ResponseEntity<Object> getUsersByUsername(@PathVariable("username") String username) {
-        List<User> users = userService.getUserUsernameStartsWith(username);
-        return new ResponseEntity<>(users, HttpStatus.OK);
+    @GetMapping(value="/user/{username}")
+    public  ResponseEntity<Object> getUser(@PathVariable("username") String username, Principal principal) {
+        if(principal.getName().equals(username)){
+        return new ResponseEntity<>(userService.getUser(username), HttpStatus.OK);}
+        else throw new NotAuthorizedException("You are not authorized to make this request");
     }
-    @GetMapping(value="/users/{username}")
-    public  ResponseEntity<Object> getUser(@PathVariable("username") String username) {
-        return new ResponseEntity<>(userService.getUser(username), HttpStatus.OK);
-    }
-    @GetMapping(value="/users/{username}/authorities")
+    @GetMapping(value="/authorities/{username}")
     public ResponseEntity<Object> getUserAuthorities(@PathVariable("username") String username) {
         return ResponseEntity.ok().body(userService.getAuthorities(username));
     }
-    @PostMapping(value= "/users")
-    public ResponseEntity<Object> createUser(@RequestBody User user) {
-        userService.save(user);
-        return new ResponseEntity<>("Account Created", HttpStatus.CREATED);
-    }
-    @PostMapping(value = "users/{username}/authorities")
+    @PostMapping(value = "authorities/{username}")
     public ResponseEntity<Object> addUserAuthority(@PathVariable("username") String username, @RequestBody Map<String, Object> fields) {
         try {
             String authorityName = (String) fields.get("authority");
@@ -58,21 +53,26 @@ public class UserController {
             throw new RecordNotFoundException();
         }
     }
-    @PutMapping(value = "users/{username}")
-    public ResponseEntity<Object> updateUser(@PathVariable("username") String username, @RequestBody User user) {
+    @PutMapping(value = "updateUser/{username}")
+    public ResponseEntity<Object> updateUser(@PathVariable("username") String username, @RequestBody User user, Principal principal) {
+        if(principal.getName().equals(username)){
         userService.updateUser(username, user);
-        return ResponseEntity.noContent().build();
+        return ResponseEntity.noContent().build();}
+        else throw new NotAuthorizedException("You are not authorized to make this request");
     }
-    @PutMapping(value = "users/{username}/changepassword")
-    public ResponseEntity<Object> updateUserPassword(@PathVariable("username") String username, @RequestBody User user) {
-        userService.updateUserPassword(username, user);
-        return ResponseEntity.noContent().build();
+    @PutMapping(value = "/changepassword/{username}")
+    public ResponseEntity<Object> updateUserPassword(@PathVariable("username") String username, @RequestBody User user, Principal principal) {
+        if(principal.getName().equals(username)){
+            userService.updateUserPassword(username, user);
+            return ResponseEntity.noContent().build();}
+        else throw new NotAuthorizedException("You are not authorized to make this request");
     }
-
-    @DeleteMapping(value="/users/{username}")
-    public ResponseEntity<Object> deleteUser(@PathVariable("username") String username) {
-        userService.deleteById(username);
-        return new ResponseEntity<>("Account deleted", HttpStatus.OK);
+    @DeleteMapping(value="/deleteUser/{username}")
+    public ResponseEntity<Object> deleteUser(@PathVariable("username") String username, Principal principal) {
+        if(principal.getName().equals(username)){
+            userService.deleteById(username);
+            return new ResponseEntity<>("Account deleted", HttpStatus.OK);}
+        else throw new NotAuthorizedException("You are not authorized to make this request");
     }
 }
